explizit casten

+++ uhr.ino
@@ -14,12 +14,12 @@ bool getNTPtime(int sec) {
   //strftime(time_output, 30, "%a  %d-%m-%y %T", localtime(&now));
   //Serial.println(time_output);
   //Serial.println();
-  sekunden.einer  = timeinfo.tm_sec  % 10;
-  sekunden.zehner = timeinfo.tm_sec  / 10;
-  minuten.einer   = timeinfo.tm_min  % 10;
-  minuten.zehner  = timeinfo.tm_min  / 10;
-  stunden.einer   = timeinfo.tm_hour % 10;
-  stunden.zehner  = timeinfo.tm_hour / 10;
+  sekunden.einer  = (unsigned short)timeinfo.tm_sec  % 10;
+  sekunden.zehner = (unsigned short)timeinfo.tm_sec  / 10;
+  minuten.einer   = (unsigned short)timeinfo.tm_min  % 10;
+  minuten.zehner  = (unsigned short)timeinfo.tm_min  / 10;
+  stunden.einer   = (unsigned short)timeinfo.tm_hour % 10;
+  stunden.zehner  = (unsigned short)timeinfo.tm_hour / 10;
 #ifdef DEBUG
   // Uhr Ausgabe:
   Serial.print(stunden.zehner);
